Moved arg parsing and config dictionary updating into it's own file
